requirements.txt
functions-framework==3.*

google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-api-python-client==2.101.0